Something in the react tool chain became unhappy when the impl and abstrat class were in different files. So put them in one file.
